requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
psycopg2-binary
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
redis